+++ src/suou/sqlalchemy/__init__.py
@@ -113,7 +113,7 @@ class AuthSrc(metaclass=ABCMeta):
         pass
 
 
-@deprecated('not working and too complex to use')
+@deprecated('not working and too complex to use. Will be removed in 0.9.0')
 def require_auth_base(cls: type[DeclarativeBase], *, src: AuthSrc, column: str | Column[_T] = 'id', dest: str = 'user',
         required: bool = False, signed: bool = False, sig_dest: str = 'signature', validators: Callable | Iterable[Callable] | None = None):
     '''
@@ -161,7 +161,7 @@ def require_auth_base(cls: type[DeclarativeBase], *, src: AuthSrc, column: str |
 from .asyncio import SQLAlchemy, AsyncSelectPagination, async_query
 from .orm import (
     id_column, snowflake_column, match_column, match_constraint, bool_column, declarative_base, parent_children,
-    author_pair, age_pair, bound_fk, unbound_fk, want_column, a_relationship, BitSelector, secret_column
+    author_pair, age_pair, bound_fk, unbound_fk, want_column, a_relationship, BitSelector, secret_column, username_column
 )
 
 # Optional dependency: do not import into __init__.py
@@ -169,7 +169,7 @@ __all__ = (
     'IdType', 'id_column', 'snowflake_column', 'entity_base', 'declarative_base', 'token_signer',
     'match_column', 'match_constraint', 'bool_column', 'parent_children',
     'author_pair', 'age_pair', 'bound_fk', 'unbound_fk', 'want_column',
-    'a_relationship', 'BitSelector', 'secret_column',
+    'a_relationship', 'BitSelector', 'secret_column', 'username_column',
     # .asyncio
     'SQLAlchemy', 'AsyncSelectPagination', 'async_query', 'SessionWrapper'
 )